requirements.txt
git+https://github.com/nandantumu/f1tenth_gym.git#egg=f110_gym
numpy
matplotlib
seaborn
scipy
torch
joblib
tqdm
tensorboardX
pandas
joblib
pillow
sklearn
pyyaml
av2
ipywidgets
pyyaml
Shapely